edited enemyMovement.cs to have direction vector, spiderMovement in progress, cleaned up knockBack

+++ pc5game/Assets/Scripts/CharacterStats.cs
@@ -55,7 +55,7 @@ public class CharacterStats : MonoBehaviour
         return speed;
     }
 
-    public void setSpeed(int newSpeed)
+    public void setSpeed(float newSpeed)
     {
         speed = newSpeed;
     }
